backend/pyproject.toml
[tool.poetry]
name = "your-virtual-env"
version = "0.1.0"
description = "A PWA designed to demonstrate financial fraud detection using advanced capabilities in MongoDB Atlas—particularly vector search"
authors = ["Mehar Grewal <mehar.grewal@mongodb.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
boto3 = "^1.35.70"
botocore = "^1.35.70"
sentence-transformers = "^2.7.0"
torch = "^2.2.2"
pydantic = "^2.7.0"
uuid = "^1.30"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
